auth/requirements.txt
uvicorn
websockets==15.0.1
pyjwt
apscheduler
fastapi
fastapi[standard]
sqlalchemy
pydantic==2.11.5
pydantic_core==2.33.2
passlib==1.7.4
python-multipart
email_validator
pytz
python-jose==3.5.0
aiogram
bcrypt==4.3.0
psycopg2-binary
redis
python-dotenv
utils
python-multipart